feat: migrate LibraryTabs.spec
